Static Testing
Overview
In this assignment, you’ll conduct a dependency check. A dependency check is a type of static testing that detects vulnerabilities associated with library dependencies needed for the application. Static testing allows you to identify vulnerabilities in the code without executing the code. In this assignment, you’ll do the following actions:

Identify software security vulnerabilities by running code through a static tester.
Identify potential mitigation techniques that have been used to mitigate vulnerabilities associated with known exploits.
Scenario
You’re a senior software developer on a team of software developers. The team is responsible for a large web application that uses the Spring framework.

The software development team discussed the vulnerabilities in the code base from your manual code review and plans to mitigate them. The team also supports a new functionality that requires the addition of a new library. A best practice for making certain the code is secure is to use a dependency check to check the refactored code base and the additional library. There are tools to help with dependency checks. You’ll integrate a dependency-check tool into your vulnerability assessment workflow.

Directions
To begin, open the Module Two Coding Assignment Code Base in Eclipse. Refer to the Uploading Files to Eclipse Desktop Version Tutorial for testing the code base in Eclipse. You can find the code base and tutorial in the Supporting Materials section.

Note: Integrating the static testing tool was a non-graded task that you should have completed in Module One. You may have already completed these steps.

Review this module’s Resources section and the supporting materials linked below to complete this assignment. After you have opened the code base, integrate the Maven dependency-check plug-in for the code base. Follow the instructions in the Integrating the Maven Dependency-Check Plug-In Tutorial linked in Supporting Materials to learn how to integrate and run the dependency-check plug-in into Maven for conducting static testing. Use the instructions in the tutorial to identify the software security vulnerabilities and document them in the Module Two Coding Assignment Template. The template is linked in the What to Submit section.

Specifically, you must address the following rubric criteria:

Run the dependency check on the code base. Include a screenshot of the resulting HTML report in your Module Two Coding Assignment Template. Make certain the screenshot includes the scan information at the top of the dependency-check report.
Document the results from the dependency check. In your Module Two Coding Assignment Template, make certain to include the codes and descriptions of each dependency that you found.
Analyze the results to identify the best solutions for addressing dependencies in the code base. Summarize your findings in your Module Two Coding Assignment Template. You can refer to industry standard guidelines, such as the Common Vulnerabilities and Exposures (CVE) and the National Vulnerability Database (NVD), by reviewing the resources in this module’s Resources section.
Also consider why you should filter false positives from the dependency-check tool
Discuss this in the Module Two Coding Assignment Template.